==> spatial_HYP/code/11-GSEA of markers and sex DEGs/00-Join enrichr TF-target sets and reformat for fgsea.Rmd <==
---
title: "00-Join enrichr TF-target sets and reformat for fgsea"
author: "Bernard Mulvey"
date: "2023-12-18"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.height = 10,fig.width = 7,include = FALSE)
knitr::opts_chunk$set(fig.width=7,fig.height=10)
#### sets tab autocompletion for directories to begin from the directory containing the .Rproj session, instead of the script's directory if different
knitr::opts_knit$set(root.dir = here::here())

library(data.table)
library(Biostrings)
library(stringr)

## change write.table to output a TSV without quotes or rownames by default
library(default)
default(write.table) <- list(row.names=F,col.names=T,sep='\t',quote=F)

## enable forked parallel processing with BiocParallel::multicoreParam, future::, etc
library(parallelly)
options(parallelly.supportsMulticore.disableOn="")
options(parallelly.fork.enable=TRUE)
library(BiocParallel)
options(bphost="localhost")
```

```{r}
tftargsets <- list.files("processed-data/11-GSEA/00-enrichr TF-target sets from dbs and lit mining DLed 120623",pattern=".txt",full.names=T)

sourcename <- gsub(tftargsets,pattern="^.*DLed 120623/(.*).txt$",replacement="\\1")

# read each file as a single column table of longass lines, split at the first \t to get set name and set members separated, then expand them out to a one-to-one table to feed into split
tftargsets <- mapply(X=tftargsets,Y=sourcename,SIMPLIFY=F,FUN=function(X,Y){
  tmp <- fread(X,sep = '\n')
  tmp[,setmembers:=apply(.SD,MARGIN=1,FUN=function(x){stringi::stri_split_fixed(x,pattern="\t",n=2)[[1]][2]}),.SDcols=1]
  tmp[,setname:=apply(.SD,MARGIN=1,FUN=function(x){stringi::stri_split_fixed(x,pattern="\t",n=2)[[1]][1]}),.SDcols=1]
  tmp[,setmembers:=gsub(setmembers,pattern="\t",replacement=";")]
  tmp[,setmembers:=gsub("^;","",setmembers)]
  tmp[,setname:=gsub(setname,pattern=" ",replacement="_")]
  tmp[,sourceset:=gsub(Y,pattern=" ",replacement="_")]
  return(tmp[,c(3,2,4)])
})

# for the TF-LOF_Expression_from_GEO we need to take an extra step, since these also have scores (DE values?) with them
tftargmanual <- fread("processed-data/11-GSEA/00-enrichr TF-target sets from dbs and lit mining DLed 120623/TF-LOF_Expression_from_GEO.txt",sep="\n")
setnames(tftargmanual,1,"origstr")
tftargmanual[,setname:=apply(.SD,MARGIN=1,FUN=function(x){str_split_fixed(x,pattern="\t",n=2)[1]}),.SDcols=1]
tftargmanual[,setname:=gsub(" ","_",setname)]
tftargmanual[,setmembers:=apply(.SD,MARGIN=1,FUN=function(x){str_split_fixed(x,pattern="\t",n=2)[2]}),.SDcols=1]
tftargmanual[,setmembers2 := gsub("\\d","",setmembers)]
tftargmanual[,setmembers2 := gsub("\\.","",setmembers2)]
tftargmanual[,setmembers2:=gsub(",\t",";",setmembers2)]
tftargmanual[,setmembers2:=gsub("\t","",setmembers2)]
tftargmanual[,setmembers2:=gsub(" ","",setmembers2)]
tftargmanual[,setmembers:=setmembers2]
tftargmanual[,setmembers2:=NULL]
tftargmanual[,sourceset:="TF-LOF_Expression_from_GEO"]

### replace the original TF-LOF_Expression_from_GEO with the new one
tftargsets[[4]] <- copy(tftargmanual)[,c(3,2,4)]

# clean up
rm(tftargmanual)
```

divy into fgsea-style gene sets
```{r}
library(tidyr)
tftargsets <- lapply(tftargsets,function(x){
  x %>% separate_rows(setmembers,sep=";")
})
tftargsets <- lapply(tftargsets,function(x){
  x <- as.data.table(x)[setmembers!=" "]
  x[,set_n_source:=paste0(setname,"_",sourceset)]
})

alltfsets <- rbindlist(tftargsets,use.names=T)
alltfsets <- alltfsets[!(setmembers %in% c(" ","","\t"))]
alltfsets[,setmembers:=toupper(setmembers)]
fgseatfsets <- split(alltfsets$setmembers,alltfsets$set_n_source)

# save
saveRDS(fgseatfsets,"processed-data/11-GSEA/00a-enrichrTFsets_forfgseainput.RDS")
```

```{r}
sessionInfo()
sessioninfo::session_info()
```
R version 4.3.1 (2023-06-16)
Platform: aarch64-apple-darwin20 (64-bit)
Running under: macOS Ventura 13.6

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib 
LAPACK: /Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.11.0

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: America/Chicago
tzcode source: internal

attached base packages:
[1] stats4    stats     graphics  grDevices utils     datasets  methods  
[8] base     

other attached packages:
 [1] tidyr_1.3.0         BiocParallel_1.36.0 parallelly_1.36.0  
 [4] default_1.0.0       stringr_1.5.1       Biostrings_2.70.1  
 [7] GenomeInfoDb_1.38.2 XVector_0.42.0      IRanges_2.36.0     
[10] S4Vectors_0.40.2    BiocGenerics_0.48.1 data.table_1.14.10 
[13] rlang_1.1.2        

loaded via a namespace (and not attached):
  [1] reactable_0.4.4               matrixStats_1.2.0            
  [3] bitops_1.0-7                  webshot_0.5.5                
  [5] httr_1.4.7                    RColorBrewer_1.1-3           
  [7] doParallel_1.0.17             Rgraphviz_2.46.0             
  [9] numDeriv_2016.8-1.1           tools_4.3.1                  
 [11] backports_1.4.1               utf8_1.2.4                   
 [13] R6_2.5.1                      DT_0.31                      
 [15] lazyeval_0.2.2                withr_2.5.2                  
 [17] gridExtra_2.3                 prettyunits_1.2.0            
 [19] cli_3.6.2                     Biobase_2.62.0               
 [21] sass_0.4.8                    KEGGgraph_1.62.0             
 [23] SQUAREM_2021.1                mvtnorm_1.2-4                
 [25] dreamlet_1.0.0                mixsqp_0.3-48                
 [27] systemfonts_1.0.5             Rsamtools_2.18.0             
 [29] svglite_2.1.3                 zenith_1.4.2                 
 [31] scater_1.30.1                 sessioninfo_1.2.2            
 [33] attempt_0.3.1                 invgamma_1.1                 
 [35] maps_3.4.2                    limma_3.58.1                 
 [37] rstudioapi_0.15.0             RSQLite_2.3.4                
 [39] visNetwork_2.1.2              generics_0.1.3               
 [41] BiocIO_1.12.0                 gtools_3.9.5                 
 [43] zip_2.3.0                     dplyr_1.1.4                  
 [45] Matrix_1.6-4                  ggbeeswarm_0.7.2             
 [47] fansi_1.0.6                   abind_1.4-5                  
 [49] lifecycle_1.0.4               scatterplot3d_0.3-44         
 [51] yaml_2.3.8                    edgeR_4.0.3                  
 [53] SummarizedExperiment_1.32.0   gplots_3.1.3                 
 [55] SparseArray_1.2.2             BiocFileCache_2.10.1         
 [57] paletteer_1.5.0               grid_4.3.1                   
 [59] blob_1.2.4                    promises_1.2.1               
 [61] ExperimentHub_2.10.0          crayon_1.5.2                 
 [63] lattice_0.22-5                beachmat_2.18.0              
 [65] msigdbr_7.5.1                 cowplot_1.1.2                
 [67] annotate_1.80.0               KEGGREST_1.42.0              
 [69] magick_2.8.1                  pillar_1.9.0                 
 [71] knitr_1.45                    fgsea_1.28.0                 
 [73] GenomicRanges_1.54.1          rjson_0.2.21                 
 [75] boot_1.3-28.1                 corpcor_1.6.10               
 [77] codetools_0.2-19              fastmatch_1.1-4              
 [79] glue_1.6.2                    vctrs_0.6.5                  
 [81] png_0.1-8                     spam_2.10-0                  
 [83] Rdpack_2.6                    gtable_0.3.4                 
 [85] assertthat_0.2.1              rematch2_2.1.2               
 [87] cachem_1.0.8                  openxlsx_4.2.5.2             
 [89] xfun_0.41                     rbibutils_2.2.16             
 [91] S4Arrays_1.2.0                mime_0.12                    
 [93] Rfast_2.1.0                   SingleCellExperiment_1.24.0  
 [95] iterators_1.0.14              fields_15.2                  
 [97] statmod_1.5.0                 interactiveDisplayBase_1.40.0
 [99] ellipsis_0.3.2                nlme_3.1-164                 
[101] pbkrtest_0.5.2                bit64_4.0.5                  
[103] progress_1.2.3                EnvStats_2.8.1               
[105] filelock_1.0.3                rprojroot_2.0.4              
[107] bslib_0.6.1                   irlba_2.3.5.1                
[109] vipor_0.4.7                   KernSmooth_2.23-22           
[111] colorspace_2.1-0              rmeta_3.0                    
[113] DBI_1.1.3                     tidyselect_1.2.0             
[115] bit_4.0.5                     compiler_4.3.1               
[117] curl_5.2.0                    rvest_1.0.3                  
[119] graph_1.80.0                  BiocNeighbors_1.20.0         
[121] xml2_1.3.6                    DelayedArray_0.28.0          
[123] plotly_4.10.3                 rtracklayer_1.62.0           
[125] scales_1.3.0                  caTools_1.18.2               
[127] remaCor_0.0.16                rappdirs_0.3.3               
[129] SpatialExperiment_1.12.0      digest_0.6.33                
[131] minqa_1.2.6                   variancePartition_1.33.0     
[133] rmarkdown_2.25                benchmarkmeData_1.0.4        
[135] aod_1.3.3                     RhpcBLASctl_0.23-42          
[137] htmltools_0.5.7               pkgconfig_2.0.3              
[139] lme4_1.1-35.1                 sparseMatrixStats_1.14.0     
[141] MatrixGenerics_1.14.0         mashr_0.2.79                 
[143] dbplyr_2.4.0                  fastmap_1.1.1                
[145] htmlwidgets_1.6.4             spatialLIBD_1.14.1           
[147] shiny_1.8.0                   DelayedMatrixStats_1.24.0    
[149] farver_2.1.1                  jquerylib_0.1.4              
[151] jsonlite_1.8.8                config_0.3.2                 
[153] BiocSingular_1.18.0           RCurl_1.98-1.13              
[155] magrittr_2.0.3                kableExtra_1.3.4             
[157] scuttle_1.12.0                GenomeInfoDbData_1.2.11      
[159] dotCall64_1.1-1               munsell_0.5.0                
[161] Rcpp_1.0.11                   viridis_0.6.4                
[163] babelgene_22.9                EnrichmentBrowser_2.32.0     
[165] RcppZiggurat_0.1.6            stringi_1.8.3                
[167] zlibbioc_1.48.0               MASS_7.3-60                  
[169] AnnotationHub_3.10.0          plyr_1.8.9                   
[171] parallel_4.3.1                ggrepel_0.9.4                
[173] splines_4.3.1                 hms_1.1.3                    
[175] locfit_1.5-9.8                igraph_1.6.0                 
[177] ScaledMatrix_1.10.0           reshape2_1.4.4               
[179] BiocVersion_3.18.1            XML_3.99-0.16                
[181] evaluate_0.23                 golem_0.4.1                  
[183] RcppParallel_5.1.7            BiocManager_1.30.22          
[185] tweenr_2.0.2                  nloptr_2.0.3                 
[187] foreach_1.5.2                 httpuv_1.6.13                
[189] polyclip_1.10-6               purrr_1.0.2                  
[191] benchmarkme_1.0.8             scattermore_1.2              
[193] ggplot2_3.4.4                 ashr_2.2-63                  
[195] ggforce_0.4.1                 rsvd_1.0.5                   
[197] broom_1.0.5                   xtable_1.8-4                 
[199] restfulr_0.0.15               fANCOVA_0.6-1                
[201] later_1.3.2                   viridisLite_0.4.2            
[203] Polychrome_1.5.1              truncnorm_1.0-9              
[205] tibble_3.2.1                  hypeR_2.0.0                  
[207] lmerTest_3.1-3                beeswarm_0.4.0               
[209] memoise_2.0.1                 AnnotationDbi_1.64.1         
[211] GenomicAlignments_1.38.0      shinyWidgets_0.8.0           
[213] GSEABase_1.64.0               here_1.0.1                   
> sessioninfo::session_info()
─ Session info ────────────────────────────────────────────────────────────
 setting  value
 version  R version 4.3.1 (2023-06-16)
 os       macOS Ventura 13.6
 system   aarch64, darwin20
 ui       RStudio
 language (EN)
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       America/Chicago
 date     2023-12-18
 rstudio  2023.09.1+494 Desert Sunflower (desktop)
 pandoc   3.1.1 @ /Applications/RStudio.app/Contents/Resources/app/quarto/bin/tools/ (via rmarkdown)

─ Packages ────────────────────────────────────────────────────────────────
 package                * version    date (UTC) lib source
 abind                    1.4-5      2016-07-21 [1] CRAN (R 4.3.0)
 annotate                 1.80.0     2023-10-24 [1] Bioconductor
 AnnotationDbi            1.64.1     2023-11-02 [1] Bioconductor
 AnnotationHub            3.10.0     2023-10-24 [1] Bioconductor
 aod                      1.3.3      2023-12-13 [1] CRAN (R 4.3.1)
 ashr                     2.2-63     2023-08-21 [1] CRAN (R 4.3.0)
 assertthat               0.2.1      2019-03-21 [1] CRAN (R 4.3.0)
 attempt                  0.3.1      2020-05-03 [1] CRAN (R 4.3.0)
 babelgene                22.9       2022-09-29 [1] CRAN (R 4.3.0)
 backports                1.4.1      2021-12-13 [1] CRAN (R 4.3.0)
 beachmat                 2.18.0     2023-10-24 [1] Bioconductor
 beeswarm                 0.4.0      2021-06-01 [1] CRAN (R 4.3.0)
 benchmarkme              1.0.8      2022-06-12 [1] CRAN (R 4.3.0)
 benchmarkmeData          1.0.4      2020-04-23 [1] CRAN (R 4.3.0)
 Biobase                  2.62.0     2023-10-26 [1] Bioconductor
 BiocFileCache            2.10.1     2023-10-26 [1] Bioconductor
 BiocGenerics           * 0.48.1     2023-11-02 [1] Bioconductor
 BiocIO                   1.12.0     2023-10-24 [1] Bioconductor
 BiocManager              1.30.22    2023-08-08 [1] CRAN (R 4.3.0)
 BiocNeighbors            1.20.0     2023-10-24 [1] Bioconductor
 BiocParallel           * 1.36.0     2023-10-26 [1] Bioconductor
 BiocSingular             1.18.0     2023-10-24 [1] Bioconductor
 BiocVersion              3.18.1     2023-11-15 [1] Bioconductor
 Biostrings             * 2.70.1     2023-10-26 [1] Bioconductor
 bit                      4.0.5      2022-11-15 [1] CRAN (R 4.3.0)
 bit64                    4.0.5      2020-08-30 [1] CRAN (R 4.3.0)
 bitops                   1.0-7      2021-04-24 [1] CRAN (R 4.3.0)
 blob                     1.2.4      2023-03-17 [1] CRAN (R 4.3.0)
 boot                     1.3-28.1   2022-11-22 [1] CRAN (R 4.3.0)
 broom                    1.0.5      2023-06-09 [1] CRAN (R 4.3.0)
 bslib                    0.6.1      2023-11-28 [1] CRAN (R 4.3.1)
 cachem                   1.0.8      2023-05-01 [1] CRAN (R 4.3.0)
 caTools                  1.18.2     2021-03-28 [1] CRAN (R 4.3.0)
 cli                      3.6.2      2023-12-11 [1] CRAN (R 4.3.1)
 codetools                0.2-19     2023-02-01 [1] CRAN (R 4.3.0)
 colorspace               2.1-0      2023-01-23 [1] CRAN (R 4.3.0)
 config                   0.3.2      2023-08-30 [1] CRAN (R 4.3.0)
 corpcor                  1.6.10     2021-09-16 [1] CRAN (R 4.3.0)
 cowplot                  1.1.2      2023-12-15 [1] CRAN (R 4.3.1)
 crayon                   1.5.2      2022-09-29 [1] CRAN (R 4.3.0)
 curl                     5.2.0      2023-12-08 [1] CRAN (R 4.3.1)
 data.table             * 1.14.10    2023-12-08 [1] CRAN (R 4.3.1)
 DBI                      1.1.3      2022-06-18 [1] CRAN (R 4.3.0)
 dbplyr                   2.4.0      2023-10-26 [1] CRAN (R 4.3.1)
 default                * 1.0.0      2017-08-07 [1] CRAN (R 4.3.0)
 DelayedArray             0.28.0     2023-11-06 [1] Bioconductor
 DelayedMatrixStats       1.24.0     2023-10-24 [1] Bioconductor
 digest                   0.6.33     2023-07-07 [1] CRAN (R 4.3.0)
 doParallel               1.0.17     2022-02-07 [1] CRAN (R 4.3.0)
 dotCall64                1.1-1      2023-11-28 [1] CRAN (R 4.3.1)
 dplyr                    1.1.4      2023-11-17 [1] CRAN (R 4.3.1)
 dreamlet                 1.0.0      2023-11-06 [1] Bioconductor
 DT                       0.31       2023-12-09 [1] CRAN (R 4.3.1)
 edgeR                    4.0.3      2023-12-09 [1] Bioconductor 3.18 (R 4.3.2)
 ellipsis                 0.3.2      2021-04-29 [1] CRAN (R 4.3.0)
 EnrichmentBrowser        2.32.0     2023-10-24 [1] Bioconductor
 EnvStats                 2.8.1      2023-08-22 [1] CRAN (R 4.3.0)
 evaluate                 0.23       2023-11-01 [1] CRAN (R 4.3.1)
 ExperimentHub            2.10.0     2023-10-24 [1] Bioconductor
 fANCOVA                  0.6-1      2020-11-13 [1] CRAN (R 4.3.1)
 fansi                    1.0.6      2023-12-08 [1] CRAN (R 4.3.1)
 farver                   2.1.1      2022-07-06 [1] CRAN (R 4.3.0)
 fastmap                  1.1.1      2023-02-24 [1] CRAN (R 4.3.0)
 fastmatch                1.1-4      2023-08-18 [1] CRAN (R 4.3.0)
 fgsea                    1.28.0     2023-10-24 [1] Bioconductor
 fields                   15.2       2023-08-17 [1] CRAN (R 4.3.0)
 filelock                 1.0.3      2023-12-11 [1] CRAN (R 4.3.1)
 foreach                  1.5.2      2022-02-02 [1] CRAN (R 4.3.0)
 generics                 0.1.3      2022-07-05 [1] CRAN (R 4.3.0)
 GenomeInfoDb           * 1.38.2     2023-12-13 [1] Bioconductor 3.18 (R 4.3.1)
 GenomeInfoDbData         1.2.11     2023-11-10 [1] Bioconductor
 GenomicAlignments        1.38.0     2023-10-24 [1] Bioconductor
 GenomicRanges            1.54.1     2023-10-30 [1] Bioconductor
 ggbeeswarm               0.7.2      2023-04-29 [1] CRAN (R 4.3.0)
 ggforce                  0.4.1      2022-10-04 [1] CRAN (R 4.3.0)
 ggplot2                  3.4.4      2023-10-12 [1] CRAN (R 4.3.1)
 ggrepel                  0.9.4      2023-10-13 [1] CRAN (R 4.3.1)
 glue                     1.6.2      2022-02-24 [1] CRAN (R 4.3.0)
 golem                    0.4.1      2023-06-05 [1] CRAN (R 4.3.0)
 gplots                   3.1.3      2022-04-25 [1] CRAN (R 4.3.0)
 graph                    1.80.0     2023-10-26 [1] Bioconductor
 gridExtra                2.3        2017-09-09 [1] CRAN (R 4.3.0)
 GSEABase                 1.64.0     2023-10-24 [1] Bioconductor
 gtable                   0.3.4      2023-08-21 [1] CRAN (R 4.3.0)
 gtools                   3.9.5      2023-11-20 [1] CRAN (R 4.3.1)
 here                     1.0.1      2020-12-13 [1] CRAN (R 4.3.0)
 hms                      1.1.3      2023-03-21 [1] CRAN (R 4.3.0)
 htmltools                0.5.7      2023-11-03 [1] CRAN (R 4.3.1)
 htmlwidgets              1.6.4      2023-12-06 [1] CRAN (R 4.3.1)
 httpuv                   1.6.13     2023-12-06 [1] CRAN (R 4.3.1)
 httr                     1.4.7      2023-08-15 [1] CRAN (R 4.3.0)
 hypeR                    2.0.0      2023-10-24 [1] Bioconductor
 igraph                   1.6.0      2023-12-11 [1] CRAN (R 4.3.1)
 interactiveDisplayBase   1.40.0     2023-10-24 [1] Bioconductor
 invgamma                 1.1        2017-05-07 [1] CRAN (R 4.3.0)
 IRanges                * 2.36.0     2023-10-26 [1] Bioconductor
 irlba                    2.3.5.1    2022-10-03 [1] CRAN (R 4.3.1)
 iterators                1.0.14     2022-02-05 [1] CRAN (R 4.3.0)
 jquerylib                0.1.4      2021-04-26 [1] CRAN (R 4.3.0)
 jsonlite                 1.8.8      2023-12-04 [1] CRAN (R 4.3.1)
 kableExtra               1.3.4      2021-02-20 [1] CRAN (R 4.3.0)
 KEGGgraph                1.62.0     2023-10-24 [1] Bioconductor
 KEGGREST                 1.42.0     2023-10-26 [1] Bioconductor
 KernSmooth               2.23-22    2023-07-10 [1] CRAN (R 4.3.0)
 knitr                    1.45       2023-10-30 [1] CRAN (R 4.3.1)
 later                    1.3.2      2023-12-06 [1] CRAN (R 4.3.1)
 lattice                  0.22-5     2023-10-24 [1] CRAN (R 4.3.1)
 lazyeval                 0.2.2      2019-03-15 [1] CRAN (R 4.3.0)
 lifecycle                1.0.4      2023-11-07 [1] CRAN (R 4.3.1)
 limma                    3.58.1     2023-10-31 [1] Bioconductor
 lme4                     1.1-35.1   2023-11-05 [1] CRAN (R 4.3.1)
 lmerTest                 3.1-3      2020-10-23 [1] CRAN (R 4.3.0)
 locfit                   1.5-9.8    2023-06-11 [1] CRAN (R 4.3.0)
 magick                   2.8.1      2023-10-22 [1] CRAN (R 4.3.1)
 magrittr                 2.0.3      2022-03-30 [1] CRAN (R 4.3.0)
 maps                     3.4.2      2023-12-15 [1] CRAN (R 4.3.1)
 mashr                    0.2.79     2023-10-18 [1] CRAN (R 4.3.1)
 MASS                     7.3-60     2023-05-04 [1] CRAN (R 4.3.0)
 Matrix                   1.6-4      2023-11-30 [1] CRAN (R 4.3.1)
 MatrixGenerics           1.14.0     2023-10-26 [1] Bioconductor
 matrixStats              1.2.0      2023-12-11 [1] CRAN (R 4.3.1)
 memoise                  2.0.1      2021-11-26 [1] CRAN (R 4.3.0)
 mime                     0.12       2021-09-28 [1] CRAN (R 4.3.0)
 minqa                    1.2.6      2023-09-11 [1] CRAN (R 4.3.0)
 mixsqp                   0.3-48     2022-11-16 [1] CRAN (R 4.3.0)
 msigdbr                  7.5.1      2022-03-30 [1] CRAN (R 4.3.0)
 munsell                  0.5.0      2018-06-12 [1] CRAN (R 4.3.0)
 mvtnorm                  1.2-4      2023-11-27 [1] CRAN (R 4.3.1)
 nlme                     3.1-164    2023-11-27 [1] CRAN (R 4.3.1)
 nloptr                   2.0.3      2022-05-26 [1] CRAN (R 4.3.0)
 numDeriv                 2016.8-1.1 2019-06-06 [1] CRAN (R 4.3.0)
 openxlsx                 4.2.5.2    2023-02-06 [1] CRAN (R 4.3.0)
 paletteer                1.5.0      2022-10-19 [1] CRAN (R 4.3.0)
 parallelly             * 1.36.0     2023-05-26 [1] CRAN (R 4.3.0)
 pbkrtest                 0.5.2      2023-01-19 [1] CRAN (R 4.3.0)
 pillar                   1.9.0      2023-03-22 [1] CRAN (R 4.3.0)
 pkgconfig                2.0.3      2019-09-22 [1] CRAN (R 4.3.0)
 plotly                   4.10.3     2023-10-21 [1] CRAN (R 4.3.1)
 plyr                     1.8.9      2023-10-02 [1] CRAN (R 4.3.1)
 png                      0.1-8      2022-11-29 [1] CRAN (R 4.3.0)
 Polychrome               1.5.1      2022-05-03 [1] CRAN (R 4.3.0)
 polyclip                 1.10-6     2023-09-27 [1] CRAN (R 4.3.1)
 prettyunits              1.2.0      2023-09-24 [1] CRAN (R 4.3.1)
 progress                 1.2.3      2023-12-06 [1] CRAN (R 4.3.1)
 promises                 1.2.1      2023-08-10 [1] CRAN (R 4.3.0)
 purrr                    1.0.2      2023-08-10 [1] CRAN (R 4.3.0)
 R6                       2.5.1      2021-08-19 [1] CRAN (R 4.3.0)
 rappdirs                 0.3.3      2021-01-31 [1] CRAN (R 4.3.0)
 rbibutils                2.2.16     2023-10-25 [1] CRAN (R 4.3.1)
 RColorBrewer             1.1-3      2022-04-03 [1] CRAN (R 4.3.0)
 Rcpp                     1.0.11     2023-07-06 [1] CRAN (R 4.3.1)
 RcppParallel             5.1.7      2023-02-27 [1] CRAN (R 4.3.0)
 RcppZiggurat             0.1.6      2020-10-20 [1] CRAN (R 4.3.0)
 RCurl                    1.98-1.13  2023-11-02 [1] CRAN (R 4.3.1)
 Rdpack                   2.6        2023-11-08 [1] CRAN (R 4.3.1)
 reactable                0.4.4      2023-03-12 [1] CRAN (R 4.3.0)
 remaCor                  0.0.16     2023-06-21 [1] CRAN (R 4.3.0)
 rematch2                 2.1.2      2020-05-01 [1] CRAN (R 4.3.0)
 reshape2                 1.4.4      2020-04-09 [1] CRAN (R 4.3.0)
 restfulr                 0.0.15     2022-06-16 [1] CRAN (R 4.3.0)
 Rfast                    2.1.0      2023-11-09 [1] CRAN (R 4.3.1)
 Rgraphviz                2.46.0     2023-10-26 [1] Bioconductor
 RhpcBLASctl              0.23-42    2023-02-11 [1] CRAN (R 4.3.0)
 rjson                    0.2.21     2022-01-09 [1] CRAN (R 4.3.0)
 rlang                  * 1.1.2      2023-11-04 [1] CRAN (R 4.3.1)
 rmarkdown                2.25       2023-09-18 [1] CRAN (R 4.3.1)
 rmeta                    3.0        2018-03-20 [1] CRAN (R 4.3.0)
 rprojroot                2.0.4      2023-11-05 [1] CRAN (R 4.3.1)
 Rsamtools                2.18.0     2023-10-24 [1] Bioconductor
 RSQLite                  2.3.4      2023-12-08 [1] CRAN (R 4.3.1)
 rstudioapi               0.15.0     2023-07-07 [1] CRAN (R 4.3.0)
 rsvd                     1.0.5      2021-04-16 [1] CRAN (R 4.3.0)
 rtracklayer              1.62.0     2023-10-26 [1] Bioconductor
 rvest                    1.0.3      2022-08-19 [1] CRAN (R 4.3.0)
 S4Arrays                 1.2.0      2023-10-26 [1] Bioconductor
 S4Vectors              * 0.40.2     2023-11-25 [1] Bioconductor 3.18 (R 4.3.2)
 sass                     0.4.8      2023-12-06 [1] CRAN (R 4.3.1)
 ScaledMatrix             1.10.0     2023-10-24 [1] Bioconductor
 scales                   1.3.0      2023-11-28 [1] CRAN (R 4.3.1)
 scater                   1.30.1     2023-11-16 [1] Bioconductor
 scattermore              1.2        2023-06-12 [1] CRAN (R 4.3.0)
 scatterplot3d            0.3-44     2023-05-05 [1] CRAN (R 4.3.0)
 scuttle                  1.12.0     2023-10-24 [1] Bioconductor
 sessioninfo              1.2.2      2021-12-06 [1] CRAN (R 4.3.0)
 shiny                    1.8.0      2023-11-17 [1] CRAN (R 4.3.1)
 shinyWidgets             0.8.0      2023-08-30 [1] CRAN (R 4.3.0)
 SingleCellExperiment     1.24.0     2023-11-06 [1] Bioconductor
 spam                     2.10-0     2023-10-23 [1] CRAN (R 4.3.1)
 SparseArray              1.2.2      2023-11-08 [1] Bioconductor
 sparseMatrixStats        1.14.0     2023-10-24 [1] Bioconductor
 SpatialExperiment        1.12.0     2023-10-24 [1] Bioconductor
 spatialLIBD              1.14.1     2023-11-30 [1] Bioconductor 3.18 (R 4.3.1)
 SQUAREM                  2021.1     2021-01-13 [1] CRAN (R 4.3.0)
 statmod                  1.5.0      2023-01-06 [1] CRAN (R 4.3.0)
 stringi                  1.8.3      2023-12-11 [1] CRAN (R 4.3.1)
 stringr                * 1.5.1      2023-11-14 [1] CRAN (R 4.3.1)
 SummarizedExperiment     1.32.0     2023-11-06 [1] Bioconductor
 svglite                  2.1.3      2023-12-08 [1] CRAN (R 4.3.1)
 systemfonts              1.0.5      2023-10-09 [1] CRAN (R 4.3.1)
 tibble                   3.2.1      2023-03-20 [1] CRAN (R 4.3.0)
 tidyr                  * 1.3.0      2023-01-24 [1] CRAN (R 4.3.0)
 tidyselect               1.2.0      2022-10-10 [1] CRAN (R 4.3.0)
 truncnorm                1.0-9      2023-03-20 [1] CRAN (R 4.3.0)
 tweenr                   2.0.2      2022-09-06 [1] CRAN (R 4.3.0)
 utf8                     1.2.4      2023-10-22 [1] CRAN (R 4.3.1)
 variancePartition        1.33.0     2023-11-19 [1] Github (DiseaseNeurogenomics/variancePartition@81bcb3c)
 vctrs                    0.6.5      2023-12-01 [1] CRAN (R 4.3.1)
 vipor                    0.4.7      2023-12-18 [1] CRAN (R 4.3.1)
 viridis                  0.6.4      2023-07-22 [1] CRAN (R 4.3.0)
 viridisLite              0.4.2      2023-05-02 [1] CRAN (R 4.3.0)
 visNetwork               2.1.2      2022-09-29 [1] CRAN (R 4.3.0)
 webshot                  0.5.5      2023-06-26 [1] CRAN (R 4.3.0)
 withr                    2.5.2      2023-10-30 [1] CRAN (R 4.3.1)
 xfun                     0.41       2023-11-01 [1] CRAN (R 4.3.1)
 XML                      3.99-0.16  2023-11-29 [1] CRAN (R 4.3.1)
 xml2                     1.3.6      2023-12-04 [1] CRAN (R 4.3.1)
 xtable                   1.8-4      2019-04-21 [1] CRAN (R 4.3.0)
 XVector                * 0.42.0     2023-10-26 [1] Bioconductor
 yaml                     2.3.8      2023-12-11 [1] CRAN (R 4.3.1)
 zenith                   1.4.2      2023-11-19 [1] Github (DiseaseNeurogenomics/zenith@7fd3e8e)
 zip                      2.3.0      2023-04-17 [1] CRAN (R 4.3.0)
 zlibbioc                 1.48.0     2023-10-26 [1] Bioconductor

 [1] /Users/bmulvey/Library/R/arm64/4.3/library
 [2] /Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library

───────────────────────────────────────────────────────────────────────────
