==> spatial_HYP/code/11-GSEA of markers and sex DEGs/07-Sex DEG hypergeometric tests for GWAS gene sets.Rmd <==
---
title: "07-Sex DEG hypergeometric tests for GWAS gene sets"
author: "Bernard Mulvey"
date: "2024-01-13"
output: html_document
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.height = 10,fig.width = 7,include = FALSE)
knitr::opts_chunk$set(fig.width=7,fig.height=10)
#### sets tab autocompletion for directories to begin from the directory containing the .Rproj session, instead of the script's directory if different
knitr::opts_knit$set(root.dir = here::here())

knitr::opts_knit$set(root.dir = here::here())

library(data.table)
library(Biostrings)
library(hypeR)
library(BiocParallel)
library(parallelly)


require(colorout)
ColorOut()
options("styler.addins_style_transformer" = "biocthis::bioc_style()")

```

```{r}
degs <- readRDS("processed-data/09-Sex DE/01-voomLmFit_svg10-svg20-hvg20_Hmnydflt-mnn30-HmnylmbNA_BS-15-20-31-15VMHARCclpsd.RDS")

### subset to harmony lambda=NULL nnsvg10 runs (k15 k20 k31 and k15 with vmh and arc collapsed to single domains)
degs <- degs[grep(names(degs),pattern="HARMONYlmbna_nnsvg10")]

un <- fread("raw-data/Genesets from GWAS pubs, TWAS, and disgenet/nonredundant_union_set_genesymbols.txt")

sbp <- MulticoreParam(4)
reslist <- bplapply(degs, BPPARAM=sbp,function(x){
    i <- 1
    ndom <- length(unique(x$assay))
    for (i in c(1:(2*ndom))){
        if (i<=ndom){
            d <- i
            dom <- unique(x$assay)[d]
            domname <- paste0(dom,"_MaleUp")
        }
        else{
            d <- i-ndom
            dom <- unique(x$assay)[d]
            domname <- paste0(dom,"_FemUp")
        }
        
        curde <- x[assay==dom]
        
        if (i<=ndom){
            cursig <- curde[P.Value<0.05&logFC>0,gene_name]
        }
        else {
            cursig <- curde[P.Value<0.05&logFC<0,gene_name]
        }
        
        if(length(cursig)==0){
            next(i)
        }
        

        # subset disease genes to unique, shared genes analyzed for DE in the current domain and format into a list of vectors for each disease
        j<-1
        curun <- un[symbol %in% x[assay==dom,gene_name]]
        tmplists <- list()
        for (j in c(1:length(unique(un$dx)))){
            curdx <- unique(un$dx)[j]
            curdxset <- unique(curun[dx==curdx,symbol])
            tmplists[[j]] <- curdxset
            names(tmplists)[j] <- curdx
            # belt and suspenders clearout
            rm(curdx,curdxset)
        }
        rm(j)
        
        # actual hypergeo tests
        tmpres <- hypeR::hypeR(signature=cursig,genesets=tmplists,background = nrow(curde)-length(cursig),test = "hypergeometric")
        tmpres2 <- as.data.table(tmpres$data)
        tmpres2[,spdomain_sex:=domname]
    
        if (i==1){
            currestab <- copy(tmpres2)
        }
        else{
            currestab <- rbind(copy(tmpres2),currestab)
        }
            rm(tmpres,tmpres2,curun,tmplists)
    }
    return(currestab)
})
rm(i,j)

i<-1
for (i in c(1:length(reslist))){
    if (i==1){
        resdat <- copy(reslist[[i]])
        resdat[,clus_set:=names(reslist)[i]]
    }
    else{
        tmp <- copy(reslist[[i]])
        tmp[,clus_set:=names(reslist)[i]]
        resdat <- rbind(resdat,tmp)
        rm(tmp)
    }
} 
    
fwrite(resdat,"processed-data/11-GSEA/07-hypergeometric enrichment for FDR-sig sex DEGs per harmonyLambdaNA-nnsvg10 cluster (3 k vals and collapsed vmh-arc) in neuropsych GWAS-TWAS-litmined union sets.txt",sep="\t",quote=F)
```

```{r}
sessionInfo()
sessioninfo::session_info()
```
