==> spatial_HYP/code/11-GSEA of markers and sex DEGs/09-Filter sex DE TF-targ results to expressed TFs.Rmd <==
---
title: "09-Filter sex DE TF-targ results to expressed TFs"
author: "Bernard Mulvey"
date: "2023-12-28"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.height = 10,fig.width = 7,include = FALSE)
knitr::opts_chunk$set(fig.width=7,fig.height=10)
#### sets tab autocompletion for directories to begin from the directory containing the .Rproj session, instead of the script's directory if different
knitr::opts_knit$set(root.dir = here::here())

library(data.table)
library(SpatialExperiment)
require(colorout)
ColorOut()
options("styler.addins_style_transformer" = "biocthis::bioc_style()")

## change write.table to output a TSV without quotes or rownames by default
library(default)
default(write.table) <- list(row.names=F,col.names=T,sep='\t',quote=F)

## enable forked parallel processing with BiocParallel::multicoreParam, future::, etc
library(parallelly)
options(parallelly.supportsMulticore.disableOn="")
options(parallelly.fork.enable=TRUE)
library(BiocParallel)
options(bphost="localhost")


## last of all, unload the base package datasets, whose data keep getting in the way of autocompletions (e.g., Theoph is priortized over TRUE)
unloadNamespace("datasets")
```

load Sex DE - TFtarg GSEA results before having run collapsePathways (which could have e.g., removed a result for a TF that's relevant in the data but retained one for a TF with similar targets despite the TF NOT being expressed in the data) and the SPE (which we need for filtering on expression)
```{r}
## un-redundancied results
sexreslist <- readRDS("processed-data/11-GSEA/04a-Sex DE TF-target GSEA svg10-hmnyLmdaNA_BS60k-k15-k20-k31-k15clpsd.RDS")

sexreslist.unsign <- readRDS("processed-data/11-GSEA/04d-Unsigned Sex DE TF-target GSEA svg10-hmnyLmdaNA_BS60k-k15-k20-k31-k15clpsd.RDS")

names(sexreslist.unsign) <- paste0("unsigned_",names(sexreslist.unsign))

sexreslist <- c(sexreslist,sexreslist.unsign)

# 
hyp2 <- readRDS("processed-data/03-QC_filters/hypN10_umi210_gene126_chrm50_spotsweeped_lognorm_rotsNmirrors_072224.RDS")

# set rownames of spe to gene symbol, since those are the identifiers we're stuck with for TF-targ sets
# make all upper case to be sure
dropg <- as.data.table(rowData(hyp2),keep.rownames=T)[,.(.N,gene_id),by=gene_name][N>1,gene_id]
hyp2 <- hyp2[!(rownames(hyp2)%in%dropg),]
rownames(hyp2) <- rowData(hyp2)$gene_name
rownames(hyp2) <- toupper(rownames(hyp2))

#### separate out the TF symbol from the full set name so we can filter the SPE accordingly
sexreslist <- lapply(sexreslist,FUN=function(x){
    
    y <- x[!(pathway %in% grep(pathway,pattern="Rummagene",value=T))]
    y[,tf:=apply(.SD,MARGIN = 1,FUN=function(h){strsplit(h,"_",1)[[1]][1]}),.SDcols="pathway"]
    ### rummagene TFs are named differently:
    z <- x[pathway %in% grep(pathway,pattern="Rummagene",value=T)]
    # reverse the output of a strsplit of a strsplit o.0 to get the string immediately preceding "_Rummagene_.."
    z[,tf:=apply(.SD,MARGIN=1,FUN=function(p){
        rev.default(strsplit(
            strsplit(p,"_Rummagene_transcription_factors",-1)[[1]][1],
            "_")[[1]]
        )[1]
    }),.SDcols="pathway"]
    #z[,tf:=gsub(pathway,pattern=".*-.*_(.*)_Rummagene_transcription_factors)",replacement="\\1")]
    fix <- rbind(z,y)
    # make all upper case
    fix[,tf:=toupper(tf)]
    return(fix)
})

keepgenes <- unique(unlist(lapply(sexreslist,FUN=function(x){x$tf})))
keepgenes <- keepgenes[keepgenes %in% rownames(hyp2)]
hyp2 <- hyp2[keepgenes,]

rm(keepgenes)
gc(full=T)

### finally, rearrange the results into a list of lists -- each parent list a clustering run, each child list one of the clusters
names(sexreslist)

sexreslist <- list(
    sexreslist[grep(names(sexreslist),pattern="^BSpace_k15_HARMONYlmbna_nnsvg10_X",value=T)],
    sexreslist[grep(names(sexreslist),pattern="^BSpace_k20_HARMONYlmbna_nnsvg10",value=T)],
    sexreslist[grep(names(sexreslist),pattern="^BSpace_k31_HARMONYlmbna_nnsvg10",value=T)],
    sexreslist[grep(names(sexreslist),pattern="^BSpace_k15_HARMONYlmbna_nnsvg10_clpsdARC_clpsdVMH",value=T)],
    sexreslist[grep(names(sexreslist),pattern="^unsigned_BSpace_k15_HARMONYlmbna_nnsvg10_X",value=T)],
    sexreslist[grep(names(sexreslist),pattern="^unsigned_BSpace_k20_HARMONYlmbna_nnsvg10",value=T)],
    sexreslist[grep(names(sexreslist),pattern="^unsigned_BSpace_k31_HARMONYlmbna_nnsvg10",value=T)],
    sexreslist[grep(names(sexreslist),pattern="^unsigned_BSpace_k15_HARMONYlmbna_nnsvg10_clpsdARC_clpsdVMH",value=T)]
)

names(sexreslist) <- c(
    "BSpace_k15_HARMONYlmbna_nnsvg10",
    "BSpace_k20_HARMONYlmbna_nnsvg10",
    "BSpace_k31_HARMONYlmbna_nnsvg10",
    "BSpace_k15_HARMONYlmbna_nnsvg10_clpsdARC_clpsdVMH",
    "unsigned_BSpace_k15_HARMONYlmbna_nnsvg10",
    "unsigned_BSpace_k20_HARMONYlmbna_nnsvg10",
    "unsigned_BSpace_k31_HARMONYlmbna_nnsvg10",
    "unsigned_BSpace_k15_HARMONYlmbna_nnsvg10_clpsdARC_clpsdVMH")
```

### get the clustering sets for the four Bayesspace runs we're reporting so we can extract and filter to TFs expressed in the cluster at >0.0025% of spots in ≥ 3 samples and with a mean expression in at least 1% of spots for all 4 clustering approaches
```{r}
bscl <- list.files("processed-data/06-BayesSpace/01-bayesspace60kiter_k15-20-31_out",pattern=".txt",full.names = T,recursive = F)
bscls <- lapply(bscl,fread)

names(bscls) <- gsub(bscl,pattern="^.*31_out/(.*)\\.txt$",replacement="\\1")

bscls <- bscls[names(bscls) %in% grep(names(bscls),pattern="HARMONYlmbna_nnsvg10",value=T)]

### format for spe compatability (as data frame with matched order of rownames)
bscls <- lapply(bscls,FUN=function(x){
    setnames(x,2,"label")
    x[,label:=paste0("X",label)]
    x <- DataFrame(x)
    rownames(x) <- x$rn
    x <- x[colnames(hyp2),]
    return(x)
})

## (re)create the collapsed set with the same list name as its listed in the per-cluster GSEA results list 
bscls[[4]] <- copy(bscls[["BSpace_k15_HARMONYlmbna_nnsvg10"]])
bscls[[4]] <- as.data.table(bscls[[4]])
bscls[[4]][label %in% c("X7","X12"),label:="VMH"]
bscls[[4]][label %in% c("X4","X6"),label:="ARC"]
bscls[[4]] <- DataFrame(bscls[[4]])
rownames(bscls[[4]]) <- bscls[[4]]$rn
bscls[[4]] <- bscls[[4]][colnames(hyp2),]
names(bscls)[4] <- "BSpace_k15_HARMONYlmbna_nnsvg10_clpsdARC_clpsdVMH"

# make this a list of the clusterings x2 since we need to apply our functions to the signed and unsigned result lists
bscls <- c(bscls,bscls)

names(bscls)[5:8] <- paste0("unsigned_",names(bscls)[5:8])
# sanity check
stopifnot(names(bscls)==names(sexreslist))

```

### get TFs qualifying as expressed in each cluster for at least one sex (≥3 samples of a sex with detection in at least 1% of cluster spots and overall at least 2.5% of cluster spots expressing.) see code/08 for a more marked-up version explaining how the code tabulating this works.
```{r}
sbp <- MulticoreParam(8)
register(sbp)
sexreslist.filt <- bpmapply(R=sexreslist,C=bscls,N=as.list(names(bscls)),FUN=function(R,C,N){
    
    colLabels(hyp2) <- C$label
    
    filtres <- list()
    i<-1
    for (i in c(1:length(R))){
        sexwise <- lapply(c("Male","Female"),FUN=function(S){
            cursamps <- unique(colData(hyp2[,hyp2$Sex==S])[,"sample_id"])
            curclus <- gsub(names(R)[i],pattern=paste0(N,"_(.*)$"),replacement="\\1")
            tflogcts <- as.data.table(as.data.frame(as.matrix(logcounts(hyp2[,hyp2$label==curclus&hyp2$sample_id %in% cursamps]))),keep.rownames=T)
            setnames(tflogcts,"rn","tf")
            tflogcts <- melt(tflogcts,id.vars="tf",value.name="logcounts")
            ### get sample ID from the unique spot identifiers
            tflogcts[,samp:=gsub(variable,pattern="^.*_(.*_.*)$",replacement="\\1")]
            tflogcts[,Nspot:=.N/length(unique(tf)),by="samp"]
            
            # this is a clunky approach, but the TOTAL %age of spots expressing the TF in question (ie logcounts>0) for that sample will fill in for any rows where logcounts are >0. we can then get a table of unique combinations of sample X TF x %age of spots
            tflogcts[logcounts>0,pctxpr:=.N/Nspot,by=c("tf","samp")]
            # set NAs in pctxpr to 0
            tflogcts[is.na(pctxpr),pctxpr:=0]
            # get each unique pair of tf-sample-pctxpr, where we fetch the max row value of pctxpr for a tf-sample pair (which will be >0 and = the % of spots expressing in any row where the gene was detected for that sample)
            tffilt <- tflogcts[,max(pctxpr),by=c("tf","samp")]
            # calculate mean % spots expressing dataset wide and number of samples with expression over 1%. each tf occurs once per sample, so take the mean across rows with a given TF value for samplewise means, or tabulate the number of TF occurrences with the pctxpr filter to get the number of passing rows
            setnames(tffilt,"V1","pctxpr")
            tffilt[,meanxpr:=mean(pctxpr),by="tf"]
            tffilt[pctxpr>0.01,Nxpr:=.N,by="tf"]
            # na rows returned in Nxpr will be meaningless in this result, since we just will subset to max Nxpr analogous to what we did above, resulting in a table of X tfs with their mean + spot %age and number of samples over threshold. set na values to 0 to avoid errors relating to NAs here.
            tffilt[is.na(Nxpr),Nxpr:=0]
            tffilt2 <- tffilt[,max(Nxpr),by="tf"]
            setnames(tffilt2,"V1","Nxpr")
            ## add baack the mean expression to the number of threshold-passing samples by merging parts of the old data ttable to the above result
            tffilt2 <- merge.data.table(tffilt2,unique(tffilt[,.(tf,meanxpr)],by="tf"))
            
            # FINALLY  get the TFs meeting criteria (expressed in at least 2.5% of spots of cluster for at least 3 samples from the given sex) back and filter the result table to them
            
            keeptfs <- tffilt2[Nxpr>3&meanxpr>0.025,tf]
            # return qualifying TFs for this sex
            return(keeptfs)
        })
        
        filtres[[i]] <- R[[i]][tf %in% unique(unlist(sexwise))]
        filtres[[i]][,TF_Male_Expr:=ifelse(tf %in% sexwise[[1]],yes="yes",no="no")]
        filtres[[i]][,TF_Fem_Expr:=ifelse(tf %in% sexwise[[2]],yes="yes",no="no")]
        names(filtres)[i] <- names(R)[i]
    }
    return(filtres)
})

```

### save
```{r}
saveRDS(sexreslist.filt,"processed-data/11-GSEA/09-Sex DE TF-target GSEA results from 04a filtered to TFs expressed in at least one sex at avg 2.5pct of cluster spots and at least 1pct of cluster spots in 3samps.RDS")
```

```{r}
sessionInfo()
sessioninfo::session_info()
```

R version 4.3.2 (2023-10-31)
Platform: aarch64-apple-darwin20 (64-bit)
Running under: macOS Ventura 13.6

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib 
LAPACK: /Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.11.0

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: America/Chicago
tzcode source: internal

attached base packages:
[1] stats4    stats     graphics  grDevices utils     datasets  methods  
[8] base     

other attached packages:
 [1] parallelly_1.36.0           default_1.0.0              
 [3] colorout_1.3-0.1            BiocParallel_1.36.0        
 [5] SpatialExperiment_1.12.0    SingleCellExperiment_1.24.0
 [7] SummarizedExperiment_1.32.0 Biobase_2.62.0             
 [9] GenomicRanges_1.54.1        MatrixGenerics_1.14.0      
[11] matrixStats_1.2.0           Biostrings_2.70.1          
[13] GenomeInfoDb_1.38.5         XVector_0.42.0             
[15] IRanges_2.36.0              S4Vectors_0.40.2           
[17] BiocGenerics_0.48.1         data.table_1.14.10         
[19] rlang_1.1.3                

loaded via a namespace (and not attached):
 [1] SparseArray_1.2.3       bitops_1.0-7            lattice_0.22-5         
 [4] digest_0.6.34           magrittr_2.0.3          evaluate_0.23          
 [7] grid_4.3.2              RColorBrewer_1.1-3      fastmap_1.1.1          
[10] rprojroot_2.0.4         Matrix_1.6-5            sessioninfo_1.2.2      
[13] scales_1.3.0            codetools_0.2-19        abind_1.4-5            
[16] cli_3.6.2               crayon_1.5.2            munsell_0.5.0          
[19] DelayedArray_0.28.0     yaml_2.3.8              S4Arrays_1.2.0         
[22] parallel_4.3.2          tools_4.3.2             colorspace_2.1-0       
[25] GenomeInfoDbData_1.2.11 here_1.0.1              R6_2.5.1               
[28] lifecycle_1.0.4         magick_2.8.2            zlibbioc_1.48.0        
[31] gtable_0.3.4            glue_1.7.0              Rcpp_1.0.12            
[34] xfun_0.41               rstudioapi_0.15.0       knitr_1.45             
[37] rjson_0.2.21            htmltools_0.5.7         rmarkdown_2.25         
[40] pheatmap_1.0.12         compiler_4.3.2          RCurl_1.98-1.14        
> sessioninfo::session_info()
─ Session info ────────────────────────────────────────────────────────────
 setting  value
 version  R version 4.3.2 (2023-10-31)
 os       macOS Ventura 13.6
 system   aarch64, darwin20
 ui       RStudio
 language (EN)
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       America/Chicago
 date     2024-01-28
 rstudio  2023.09.1+494 Desert Sunflower (desktop)
 pandoc   3.1.1 @ /Applications/RStudio.app/Contents/Resources/app/quarto/bin/tools/ (via rmarkdown)

─ Packages ────────────────────────────────────────────────────────────────
 package              * version   date (UTC) lib source
 abind                  1.4-5     2016-07-21 [1] CRAN (R 4.3.0)
 Biobase              * 2.62.0    2023-10-24 [1] Bioconductor
 BiocGenerics         * 0.48.1    2023-11-02 [1] Bioconductor
 BiocParallel         * 1.36.0    2023-10-26 [1] Bioconductor
 Biostrings           * 2.70.1    2023-10-25 [1] Bioconductor
 bitops                 1.0-7     2021-04-24 [1] CRAN (R 4.3.2)
 cli                    3.6.2     2023-12-11 [1] CRAN (R 4.3.2)
 codetools              0.2-19    2023-02-01 [1] CRAN (R 4.3.0)
 colorout             * 1.3-0.1   2024-01-11 [1] local
 colorspace             2.1-0     2023-01-23 [1] CRAN (R 4.3.2)
 crayon                 1.5.2     2022-09-29 [1] CRAN (R 4.3.0)
 data.table           * 1.14.10   2023-12-08 [1] CRAN (R 4.3.2)
 default              * 1.0.0     2017-08-07 [1] CRAN (R 4.3.0)
 DelayedArray           0.28.0    2023-10-24 [1] Bioconductor
 digest                 0.6.34    2024-01-11 [1] CRAN (R 4.3.2)
 evaluate               0.23      2023-11-01 [1] CRAN (R 4.3.1)
 fastmap                1.1.1     2023-02-24 [1] CRAN (R 4.3.2)
 GenomeInfoDb         * 1.38.5    2023-12-30 [1] Bioconductor 3.18 (R 4.3.2)
 GenomeInfoDbData       1.2.11    2024-01-11 [1] Bioconductor
 GenomicRanges        * 1.54.1    2023-10-29 [1] Bioconductor
 glue                   1.7.0     2024-01-09 [1] CRAN (R 4.3.2)
 gtable                 0.3.4     2023-08-21 [1] CRAN (R 4.3.0)
 here                   1.0.1     2020-12-13 [1] CRAN (R 4.3.0)
 htmltools              0.5.7     2023-11-03 [1] CRAN (R 4.3.2)
 IRanges              * 2.36.0    2023-10-24 [1] Bioconductor
 knitr                  1.45      2023-10-30 [1] CRAN (R 4.3.2)
 lattice                0.22-5    2023-10-24 [1] CRAN (R 4.3.2)
 lifecycle              1.0.4     2023-11-07 [1] CRAN (R 4.3.1)
 magick                 2.8.2     2023-12-20 [1] CRAN (R 4.3.1)
 magrittr               2.0.3     2022-03-30 [1] CRAN (R 4.3.2)
 Matrix                 1.6-5     2024-01-11 [1] CRAN (R 4.3.2)
 MatrixGenerics       * 1.14.0    2023-10-26 [1] Bioconductor
 matrixStats          * 1.2.0     2023-12-11 [1] CRAN (R 4.3.2)
 munsell                0.5.0     2018-06-12 [1] CRAN (R 4.3.0)
 parallelly           * 1.36.0    2023-05-26 [1] CRAN (R 4.3.0)
 pheatmap               1.0.12    2019-01-04 [1] CRAN (R 4.3.0)
 R6                     2.5.1     2021-08-19 [1] CRAN (R 4.3.0)
 RColorBrewer           1.1-3     2022-04-03 [1] CRAN (R 4.3.0)
 Rcpp                   1.0.12    2024-01-09 [1] CRAN (R 4.3.2)
 RCurl                  1.98-1.14 2024-01-09 [1] CRAN (R 4.3.1)
 rjson                  0.2.21    2022-01-09 [1] CRAN (R 4.3.2)
 rlang                * 1.1.3     2024-01-10 [1] CRAN (R 4.3.2)
 rmarkdown              2.25      2023-09-18 [1] CRAN (R 4.3.2)
 rprojroot              2.0.4     2023-11-05 [1] CRAN (R 4.3.1)
 rstudioapi             0.15.0    2023-07-07 [1] CRAN (R 4.3.0)
 S4Arrays               1.2.0     2023-10-24 [1] Bioconductor
 S4Vectors            * 0.40.2    2023-11-23 [1] Bioconductor 3.18 (R 4.3.2)
 scales                 1.3.0     2023-11-28 [1] CRAN (R 4.3.2)
 sessioninfo            1.2.2     2021-12-06 [1] CRAN (R 4.3.0)
 SingleCellExperiment * 1.24.0    2023-11-06 [1] Bioconductor
 SparseArray            1.2.3     2023-12-25 [1] Bioconductor 3.18 (R 4.3.2)
 SpatialExperiment    * 1.12.0    2023-10-26 [1] Bioconductor
 SummarizedExperiment * 1.32.0    2023-11-06 [1] Bioconductor
 xfun                   0.41      2023-11-01 [1] CRAN (R 4.3.2)
 XVector              * 0.42.0    2023-10-24 [1] Bioconductor
 yaml                   2.3.8     2023-12-11 [1] CRAN (R 4.3.2)
 zlibbioc               1.48.0    2023-10-26 [1] Bioconductor

 [1] /Users/bmulvey/Library/R/arm64/4.3/library
 [2] /Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library

───────────────────────────────────────────────────────────────────────────
