==> spatial_HYP/code/11-GSEA of markers and sex DEGs/10-MarkerGSEA_svg10-hmnyLmdaNA_BS60k-k15-ARC-VMH_primCilium-genes-in-SCGSv2.Rmd <==
---
title: "10-Marker GSEA for k15 VMH, ARC in gold standard set of human primary cilium-localized and functionally necessary gene sets"
author: "Bernard Mulvey"
date: "2024-07-28"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.height = 10,fig.width = 7,include = FALSE)
knitr::opts_chunk$set(fig.width=7,fig.height=10)
#### sets tab autocompletion for directories to begin from the directory containing the .Rproj session, instead of the script's directory if different
knitr::opts_knit$set(root.dir = here::here())

library(data.table)
library(fgsea)
library(msigdbr)

require(colorout)
ColorOut()
options("styler.addins_style_transformer" = "biocthis::bioc_style()")

## change write.table to output a TSV without quotes or rownames by default
library(default)
default(write.table) <- list(row.names=F,col.names=T,sep='\t',quote=F)

## enable forked parallel processing with BiocParallel::multicoreParam, future::, etc
library(parallel)
options(parallelly.supportsMulticore.disableOn="")
options(parallelly.fork.enable=TRUE)
library(parallelly)
options(bphost="localhost")
library(BiocParallel)
```

load SpatialLIBD marker enrichment stats and subset to clusterings of interest
```{r}
mks <- readRDS("processed-data/07-Marker_genes/01b-BSpace_allruns_spatialLIBD-regwrap-enrichment_full.RDS")

mks <- mks[grep(names(mks),pattern="k15_HARMONYlmbna_nnsvg10_60kiter_collapsed")]
names(mks) <- NULL
mks <- as.data.table(mks)

setnames(mks,c(53,54),c("ens","gene"))      
```

### fgsea sets retrieval
```{r}
scgs <- fread("processed-data/11-GSEA/10a-SCGSv2 first and second order genes.txt")
setnames(scgs,c("ens","gene","classif"))

# make a list object of identifiers per  gene set
# we can map these back to the msigdb table later for reference purposes by their list name (gs_name)...
sets <- split(scgs$ens,scgs$classif)
sets[[3]] <- scgs$ens
names(sets)[3] <- "all"
```

# MARKER gsea
```{r}
### drop mito genes
mks <- mks[!(gene %in% grep(gene,pattern="MT-",value=T))]

### get named vectors of marker t-stats for each cluster
sdoms <- c("XVMH","XARC")

mklist <- list()
i<-1
for (i in c(1:length(sdoms))){
    getcol <- paste0("t_stat_",sdoms[i])
    tmp <- as.data.table(mks)[,c("ens",..getcol)]
    setorderv(tmp,getcol,-1)
    ## get the stat column as a vector and give it the gene ids as names for fgsea
    ## note that since data.table (invisibly) uses list() for column data, we have to call unlist() to get a bonafide vector below
    setnames(tmp,getcol,"stat")
    tmp <- as.data.frame(tmp)
    v <- tmp$stat
    names(v) <- tmp$ens
    mklist[[i]] <- v
    rm(v)
}
rm(i)
names(mklist) <- sdoms

### loop through each domain (each run is very fast with 10 cpus in biocparallel, so just serially run 10cpu GSEA for each cluster)
sbp <- MulticoreParam(10)
register(sbp)
mkres <- mapply(X=mklist,Y=names(mklist),SIMPLIFY=FALSE,FUN=function(X,Y){
    z<-as.data.table(fgseaMultilevel(pathways=sets,stats = X,minSize = 15,maxSize = 500,eps=0.0,BPPARAM = sbp,nPermSimple = 50000))
    z[,spdom:=Y]
    return(z)
})
names(mkres) <- names(mklist)

saveRDS(mkres,"processed-data/11-GSEA/10b-Marker GSEA-SCGSv2 cilia localized and necess gene sets collapsed k15_VMH_ARC.RDS")

mkrestab <- copy(mkres[[1]])
mkrestab[,spdom:="VMH"]
mkrestab <- rbind(mkrestab,mkres[[2]])
mkrestab[c(3:4),spdom:="ARC"]
fwrite(mkrestab,"processed-data/11-GSEA/10c-Marker GSEA-SCGSv2 cilia localized and necess gene sets collapsed k15_VMH_ARC for suptab.txt",sep='\t',quote=F)
```


```{r}
sessionInfo()
sessioninfo::session_info()
```
R version 4.4.1 RC (2024-06-06 r86719)
Platform: aarch64-apple-darwin20
Running under: macOS Sonoma 14.5

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib 
LAPACK: /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.0

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: America/Chicago
tzcode source: internal

attached base packages:
[1] parallel  stats     graphics  grDevices utils     datasets  methods   base     

other attached packages:
[1] BiocParallel_1.38.0 parallelly_1.38.0   default_1.0.0       colorout_1.3-0.2   
[5] msigdbr_7.5.1       fgsea_1.30.0        data.table_1.15.4   rlang_1.1.4        

loaded via a namespace (and not attached):
 [1] Matrix_1.7-0      babelgene_22.9    gtable_0.3.5      dplyr_1.1.4      
 [5] compiler_4.4.1    tidyselect_1.2.1  Rcpp_1.0.13       scales_1.3.0     
 [9] yaml_2.3.9        fastmap_1.2.0     lattice_0.22-6    here_1.0.1       
[13] ggplot2_3.5.1     R6_2.5.1          generics_0.1.3    knitr_1.48       
[17] tibble_3.2.1      munsell_0.5.1     rprojroot_2.0.4   pillar_1.9.0     
[21] utf8_1.2.4        fastmatch_1.1-4   xfun_0.45         cli_3.6.3        
[25] magrittr_2.0.3    digest_0.6.36     grid_4.4.1        rstudioapi_0.16.0
[29] cowplot_1.1.3     lifecycle_1.0.4   vctrs_0.6.5       evaluate_0.24.0  
[33] glue_1.7.0        codetools_0.2-20  fansi_1.0.6       colorspace_2.1-0 
[37] rmarkdown_2.27    tools_4.4.1       pkgconfig_2.0.3   htmltools_0.5.8.1
> sessioninfo::session_info()
─ Session info ───────────────────────────────────────────────────────────────────
 setting  value
 version  R version 4.4.1 RC (2024-06-06 r86719)
 os       macOS Sonoma 14.5
 system   aarch64, darwin20
 ui       RStudio
 language (EN)
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       America/Chicago
 date     2024-07-28
 rstudio  2024.07.0-daily+219 Cranberry Hibiscus (desktop)
 pandoc   3.1.11 @ /Applications/RStudio.app/Contents/Resources/app/quarto/bin/tools/aarch64/ (via rmarkdown)

─ Packages ───────────────────────────────────────────────────────────────────────
 ! package      * version date (UTC) lib source
   babelgene      22.9    2022-09-29 [1] CRAN (R 4.4.0)
   BiocParallel * 1.38.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
 P cli            3.6.3   2024-06-21 [2] CRAN (R 4.4.0)
   codetools      0.2-20  2024-03-31 [1] CRAN (R 4.4.1)
   colorout     * 1.3-0.2 2024-05-01 [1] Github (jalvesaq/colorout@c6113a2)
   colorspace     2.1-0   2023-01-23 [1] CRAN (R 4.4.0)
   cowplot        1.1.3   2024-01-22 [1] CRAN (R 4.4.0)
   data.table   * 1.15.4  2024-03-30 [2] CRAN (R 4.4.0)
   default      * 1.0.0   2017-08-07 [1] CRAN (R 4.4.0)
   digest         0.6.36  2024-06-23 [1] CRAN (R 4.4.0)
   dplyr          1.1.4   2023-11-17 [1] CRAN (R 4.4.0)
   evaluate       0.24.0  2024-06-10 [1] CRAN (R 4.4.0)
   fansi          1.0.6   2023-12-08 [1] CRAN (R 4.4.0)
   fastmap        1.2.0   2024-05-15 [1] CRAN (R 4.4.0)
   fastmatch      1.1-4   2023-08-18 [1] CRAN (R 4.4.0)
   fgsea        * 1.30.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.1)
   generics       0.1.3   2022-07-05 [1] CRAN (R 4.4.0)
   ggplot2        3.5.1   2024-04-23 [1] CRAN (R 4.4.0)
   glue           1.7.0   2024-01-09 [1] CRAN (R 4.4.0)
   gtable         0.3.5   2024-04-22 [1] CRAN (R 4.4.0)
   here           1.0.1   2020-12-13 [1] CRAN (R 4.4.0)
   htmltools      0.5.8.1 2024-04-04 [1] CRAN (R 4.4.0)
   knitr          1.48    2024-07-07 [1] CRAN (R 4.4.1)
   lattice        0.22-6  2024-03-20 [1] CRAN (R 4.4.1)
   lifecycle      1.0.4   2023-11-07 [1] CRAN (R 4.4.0)
   magrittr       2.0.3   2022-03-30 [1] CRAN (R 4.4.0)
   Matrix         1.7-0   2024-04-26 [1] CRAN (R 4.4.1)
   msigdbr      * 7.5.1   2022-03-30 [1] CRAN (R 4.4.0)
   munsell        0.5.1   2024-04-01 [1] CRAN (R 4.4.0)
   parallelly   * 1.38.0  2024-07-27 [1] CRAN (R 4.4.0)
   pillar         1.9.0   2023-03-22 [1] CRAN (R 4.4.0)
   pkgconfig      2.0.3   2019-09-22 [1] CRAN (R 4.4.0)
   R6             2.5.1   2021-08-19 [1] CRAN (R 4.4.0)
   Rcpp           1.0.13  2024-07-17 [1] CRAN (R 4.4.0)
 P rlang        * 1.1.4   2024-06-04 [2] CRAN (R 4.4.1)
   rmarkdown      2.27    2024-05-17 [1] CRAN (R 4.4.0)
   rprojroot      2.0.4   2023-11-05 [1] CRAN (R 4.4.0)
   rstudioapi     0.16.0  2024-03-24 [1] CRAN (R 4.4.0)
   scales         1.3.0   2023-11-28 [1] CRAN (R 4.4.0)
   sessioninfo    1.2.2   2021-12-06 [1] CRAN (R 4.4.0)
   tibble         3.2.1   2023-03-20 [1] CRAN (R 4.4.0)
   tidyselect     1.2.1   2024-03-11 [1] CRAN (R 4.4.0)
   utf8           1.2.4   2023-10-22 [1] CRAN (R 4.4.0)
   vctrs          0.6.5   2023-12-01 [1] CRAN (R 4.4.0)
   xfun           0.45    2024-06-16 [1] CRAN (R 4.4.0)
   yaml           2.3.9   2024-07-05 [1] CRAN (R 4.4.0)

 [1] /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library
 [2] /Users/bmulvey/Library/R/arm64/4.4/library

 P ── Loaded and on-disk path mismatch.

──────────────────────────────────────────────────────────────────────────────────
