==> spatial_HYP/code/11-GSEA of markers and sex DEGs/13-Markers, Sex DE in BrainGMT.Rmd <==
---
title: "13-Markers, Sex DE in BrainGMT"
output: html_document
date: "2024-08-12"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.height = 10,fig.width = 7,include = FALSE)
knitr::opts_chunk$set(fig.width=7,fig.height=10)
#### sets tab autocompletion for directories to begin from the directory containing the .Rproj session, instead of the script's directory if different
knitr::opts_knit$set(root.dir = here::here())

library(data.table)
library(Biostrings)
library(BiocParallel)
library(parallel)
library(parallelly)
library(fgsea)
library(clusterProfiler)
library(msigdbr)

require(colorout)
ColorOut()
options("styler.addins_style_transformer" = "biocthis::bioc_style()")

## change write.table to output a TSV without quotes or rownames by default
library(default)
default(write.table) <- list(row.names=F,col.names=T,sep='\t',quote=F)

## enable forked parallel processing with BiocParallel::multicoreParam, future::, etc
options(parallelly.supportsMulticore.disableOn="")
options(parallelly.fork.enable=TRUE)
options(bphost="localhost")

unloadNamespace("datasets")
```

load SpatialLIBD marker enrichment stats and subset to clusterings of interest
```{r}
mks <- readRDS("processed-data/07-Marker_genes/01b-BSpace_allruns_spatialLIBD-regwrap-enrichment_full.RDS")

mks <- mks[grep(names(mks),pattern="HARMONYlmbna_nnsvg10_60kiter")]
# check:
# names(mks)
# [1] "k15_HARMONYlmbna_nnsvg10_60kiter"          
# [2] "k20_HARMONYlmbna_nnsvg10_60kiter"          
# [3] "k31_HARMONYlmbna_nnsvg10_60kiter"          
# [4] "k15_HARMONYlmbna_nnsvg10_60kiter_collapsed"

```


load sex DE stats, subsetted to the four clusterings of interest (nnsvg10, harmony lambda null, q=15,20,31 and q=15 collapsed VMH/ARC)

```{r}
sexde <- readRDS("processed-data/09-Sex DE/01-voomLmFit_svg10-svg20-hvg20_Hmnydflt-mnn30-HmnylmbNA_BS-15-20-31-15VMHARCclpsd.RDS")

## subset to the nnsvg10 ones
sexde <- sexde[grep(names(sexde),pattern="HARMONYlmbna_nnsvg10")]


# check:
# names(sexde)
# [1] "k15_HARMONYlmbna_nnsvg10_60kiter"          
# [2] "k20_HARMONYlmbna_nnsvg10_60kiter"          
# [3] "k31_HARMONYlmbna_nnsvg10_60kiter"          
# [4] "k15_HARMONYlmbna_nnsvg10_60kiter_clpsdARC_clpsdVMH"

```

the brainGMT sets are in a GMT file, so import using `read.gmt` from `clusterProfiler`
```{r}
brngmt <- read.gmt("raw-data/BrainGMTv2_HumanOrthologs.gmt.txt")

## split into a list of symbols per term 
sets <- split(brngmt$gene,brngmt$term)
```

# MARKER setup
```{r}
### drop mito genes
mks <- lapply(mks,FUN=function(x){x[!(gene %in% grep(gene,pattern="MT-",value=T))]})

### get named vectors of marker t-stats for each cluster in each result set (15+20+31+13)=79
mklist <- list()
i<-1
for (i in c(1:length(mks))){
    sdoms <- grep(names(mks[[i]]),pattern="logFC_",value=T)
    sdoms <- gsub(sdoms,pattern="^logFC_(.*)$",replacement="\\1")
    j<-1
    for (j in c(1:length(sdoms))){
        getcol <- paste0("t_stat_",sdoms[j])
        tmp <- copy(mks[[i]])[,c("gene",..getcol)]
        setorderv(tmp,getcol,-1)
        ## get the stat column as a vector and give it the gene ids as names for fgsea
        ## note that since data.table (invisibly) uses list() for column data, we have to call unlist() to get a bonafide vector below
        setnames(tmp,getcol,"stat")
        tmp <- as.data.frame(tmp)
        v <- tmp$stat
        names(v) <- tmp$gene
        mklist[[length(mklist)+1]] <- v
        names(mklist)[length(mklist)] <- paste0(names(mks)[i],"_",sdoms[j])
        rm(getcol,v,tmp)
    }
    rm(j,sdoms)
}
rm(i)
```

sex DE setup
```{r}
### drop mito and sex chr genes
sexde <- lapply(sexde,FUN=function(x){
    x[!(chromosome_name %in% grep(chromosome_name,pattern="MT|X|Y",value=T))]})

### get named vectors of marker t-stats for each cluster in each result set (15+20+31+13)=79 (if no clusters dropped out of sex de analysis)
sexdelist <- list()
i<-1
for (i in c(1:length(sexde))){
    sdoms <- unique(sexde[[i]]$assay)
    setnames(sexde[[i]],"t","tstat")
    j<-1
    for (j in c(1:length(sdoms))){
        tmp <- copy(sexde[[i]][assay==sdoms[j]])[,.(gene_name,tstat)]
        setorderv(tmp,"tstat",-1)
        ## get the stat column as a vector and give it the gene ids as names for fgsea
        ## note that since data.table (invisibly) uses list() for column data, its cleaner to just make the working table a data frame for this last bit
        tmp <- as.data.frame(tmp)
        v <- tmp$tstat
        names(v) <- tmp$gene_name
        sexdelist[[length(sexdelist)+1]] <- v
        names(sexdelist)[length(sexdelist)] <- paste0(names(sexde)[i],"_",sdoms[j])
        rm(v,tmp)
    }
    rm(j,sdoms)
}
rm(i)
```

## combine these into one giant input
```{r}
names(sexdelist) <- paste0("sexDE_",names(sexdelist))
names(mklist) <- paste0("mks_",names(mklist))

alltests <- c(sexdelist,mklist)

# fix XX and XARC XVMH in list names
names(alltests) <- gsub(names(alltests),pattern="XX",replacement="X")
names(alltests) <- gsub(names(alltests),pattern="XARC",replacement="ARC")
names(alltests) <- gsub(names(alltests),pattern="XVMH",replacement="VMH")
```

clean up before starting
```{r}
rm(brngmt,mks,sexde,mklist,sexdelist)
gc(full=T)
```

run fgsea without collapse pathways for all 156 sets.
```{r}
### loop through each domain (each run is very fast with 10 cpus in biocparallel, so just serially run 10cpu GSEA for each cluster)
sbp <- MulticoreParam(10)
register(sbp)
allres <- mapply(X=alltests,Y=names(alltests),SIMPLIFY=FALSE,FUN=function(X,Y){
    z<-as.data.table(fgseaMultilevel(pathways=sets,stats = X,minSize = 15,maxSize = 500,eps=0.0,BPPARAM = sbp,nPermSimple = 50000))
    z[,spdom:=Y]
    return(z)
})
names(allres) <- names(alltests)

saveRDS(allres,"processed-data/11-GSEA/13-Marker and Sex DE brainGMT GSEA svg10-hmnyLmdaNA_BS60k-k15-k20-k31-k15clpsd.RDS")
```

## make into a suptable as well
```{r}
allrestab <- rbindlist(allres)

## fix list-ness of leadingEdge
allrestab[,leadingedge2:=apply(.SD,MARGIN=1,simplify=TRUE,FUN=function(x){paste0(unlist(x),collapse=",")}),.SDcols="leadingEdge"]

allrestab[,leadingEdge:=NULL]
setnames(allrestab,"leadingedge2","leadingEdge_genes")

### separate out into analysis type, clustering paradigm, and cluster from that paradigm
allrestab[,analysistype:=gsub(spdom,pattern="^(.*)_BSpace.*$",replacement="\\1")]
allrestab[,analysistype:=gsub(analysistype,pattern="^(.*)_k.*$",replacement="\\1")]
stopifnot(length(unique(allrestab$analysistype))==2)

allrestab[,clusteringrun:=gsub(spdom,pattern="^(.*)_BSpace_(.*)_X.*$",replacement="\\1")]
allrestab[,clusteringrun:=gsub(spdom,pattern="^.*_(k.*)_X.*$",replacement="\\1")]
allrestab[clusteringrun=="sexDE_BSpace_k15_HARMONYlmbna_nnsvg10_clpsdARC_clpsdVMH_ARC",clusteringrun:="k15_HARMONYlmbna_nnsvg10_clpsdARC_clpsdVMH"]
allrestab[clusteringrun=="sexDE_BSpace_k15_HARMONYlmbna_nnsvg10_clpsdARC_clpsdVMH_VMH",clusteringrun:="k15_HARMONYlmbna_nnsvg10_clpsdARC_clpsdVMH"]
allrestab[clusteringrun=="mks_k15_HARMONYlmbna_nnsvg10_60kiter_collapsed_ARC",clusteringrun:="k15_HARMONYlmbna_nnsvg10_clpsdARC_clpsdVMH"]
allrestab[clusteringrun=="mks_k15_HARMONYlmbna_nnsvg10_60kiter_collapsed_VMH",clusteringrun:="k15_HARMONYlmbna_nnsvg10_clpsdARC_clpsdVMH"]
allrestab[,clusteringrun:=gsub(clusteringrun,pattern="_60kiter",replacement="")]
allrestab[clusteringrun=="k15_HARMONYlmbna_nnsvg10_collapsed",clusteringrun:="k15_HARMONYlmbna_nnsvg10_clpsdARC_clpsdVMH"]
stopifnot(length(unique(allrestab$clusteringrun))==4)
allrestab[,clusteringrun:=gsub(clusteringrun,pattern="^(k.*)_.*_nnsvg10$",replacement="\\1")]
allrestab[clusteringrun=="k15_HARMONYlmbna_nnsvg10_clpsdARC_clpsdVMH",clusteringrun:="k15_collapsed"]


allrestab[,cluster:=gsub(spdom,pattern="^.*_(X.*)$",replacement="\\1")]
allrestab[,cluster:=gsub(cluster,pattern="^.*_ARC$",replacement="ARC")]
allrestab[,cluster:=gsub(cluster,pattern="^.*_VMH$",replacement="VMH")]
stopifnot(length(unique(allrestab$cluster))<=33)

allrestab[,spdom:=NULL]

## make column names informative
setnames(allrestab,c("brainGMT Gene Set","P.Value","Adj.P","log2err","Enrichment Score","Normalized Enrichment Score","Geneset Size","Leading Edge Genes","Sex DE or Marker Analysis","BayesSpace Run","Domain"))

## reorder to make more readable
allrestab <- allrestab[,c(9,10,11,1,6,3,2,5,4,7,8)]

## annotate VMH/ARC clusters in other runs
allrestab[Domain=="X4"&`BayesSpace Run`=="k15",Domain:="ARC.1"]
allrestab[Domain=="X6"&`BayesSpace Run`=="k15",Domain:="ARC.2"]
allrestab[Domain=="X7"&`BayesSpace Run`=="k15",Domain:="VMH.1"]
allrestab[Domain=="X12"&`BayesSpace Run`=="k15",Domain:="VMH.2"]

### add the k20/k31 ARC and VMH cluster identifiers as used elsewhere
allrestab[Domain=="X2"&`BayesSpace Run`=="k20",Domain:="VMH20.1"]
allrestab[Domain=="X3"&`BayesSpace Run`=="k20",Domain:="VMH20.2"]
allrestab[Domain=="X1"&`BayesSpace Run`=="k20",Domain:="ARC20.1"]
allrestab[Domain=="X4"&`BayesSpace Run`=="k20",Domain:="ARC20.2"]

allrestab[Domain=="X6"&`BayesSpace Run`=="k31",Domain:="VMH31.1"]
allrestab[Domain=="X21"&`BayesSpace Run`=="k31",Domain:="VMH31.2"]
allrestab[Domain=="X25"&`BayesSpace Run`=="k31",Domain:="VMH31.3"]
allrestab[Domain=="X26"&`BayesSpace Run`=="k31",Domain:="VMH31.4"]
allrestab[Domain=="X31"&`BayesSpace Run`=="k31",Domain:="VMH31.5"]
allrestab[Domain=="X2"&`BayesSpace Run`=="k31",Domain:="ARC31.1"]
allrestab[Domain=="X4"&`BayesSpace Run`=="k31",Domain:="ARC31.2"]
allrestab[Domain=="X11"&`BayesSpace Run`=="k31",Domain:="ARC31.3"]
allrestab[Domain=="X22"&`BayesSpace Run`=="k31",Domain:="ARC31.4"]

fwrite(allrestab,"local_supptab_assembly/marker and sexDE fGSEA results using brainGMTv2.txt",sep='\t',quote=F)
```

reprod info
```{r}
sessionInfo()
sessioninfo::session_info()
```
R version 4.4.1 RC (2024-06-06 r86719)
Platform: aarch64-apple-darwin20
Running under: macOS Sonoma 14.5

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib 
LAPACK: /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.0

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: America/Chicago
tzcode source: internal

attached base packages:
[1] parallel  stats4    stats     graphics  grDevices utils     methods   base     

other attached packages:
 [1] clusterProfiler_4.12.1 default_1.0.0          colorout_1.3-0.2      
 [4] msigdbr_7.5.1          fgsea_1.30.0           parallelly_1.38.0     
 [7] BiocParallel_1.38.0    Biostrings_2.72.1      GenomeInfoDb_1.40.1   
[10] XVector_0.44.0         IRanges_2.38.1         S4Vectors_0.42.1      
[13] BiocGenerics_0.50.0    data.table_1.15.4      rlang_1.1.4           

loaded via a namespace (and not attached):
  [1] DBI_1.2.3               gson_0.1.0              shadowtext_0.1.3       
  [4] gridExtra_2.3           magrittr_2.0.3          DOSE_3.30.1            
  [7] compiler_4.4.1          RSQLite_2.3.7           png_0.1-8              
 [10] vctrs_0.6.5             reshape2_1.4.4          stringr_1.5.1          
 [13] pkgconfig_2.0.3         crayon_1.5.3            fastmap_1.2.0          
 [16] ggraph_2.2.1            utf8_1.2.4              HDO.db_0.99.1          
 [19] rmarkdown_2.27          sessioninfo_1.2.2       enrichplot_1.24.0      
 [22] UCSC.utils_1.0.0        purrr_1.0.2             bit_4.0.5              
 [25] xfun_0.45               zlibbioc_1.50.0         cachem_1.1.0           
 [28] aplot_0.2.3             jsonlite_1.8.8          blob_1.2.4             
 [31] tweenr_2.0.3            R6_2.5.1                RColorBrewer_1.1-3     
 [34] stringi_1.8.4           GOSemSim_2.30.0         Rcpp_1.0.13            
 [37] knitr_1.48              Matrix_1.7-0            splines_4.4.1          
 [40] igraph_2.0.3            tidyselect_1.2.1        viridis_0.6.5          
 [43] qvalue_2.36.0           rstudioapi_0.16.0       yaml_2.3.9             
 [46] codetools_0.2-20        lattice_0.22-6          tibble_3.2.1           
 [49] plyr_1.8.9              treeio_1.28.0           Biobase_2.64.0         
 [52] withr_3.0.0             KEGGREST_1.44.1         evaluate_0.24.0        
 [55] gridGraphics_0.5-1      scatterpie_0.2.3        polyclip_1.10-6        
 [58] ggtree_3.12.0           pillar_1.9.0            ggfun_0.1.5            
 [61] generics_0.1.3          rprojroot_2.0.4         ggplot2_3.5.1          
 [64] tidytree_0.4.6          munsell_0.5.1           scales_1.3.0           
 [67] glue_1.7.0              lazyeval_0.2.2          tools_4.4.1            
 [70] datasets_4.4.1          babelgene_22.9          fs_1.6.4               
 [73] graphlayouts_1.1.1      fastmatch_1.1-4         tidygraph_1.3.1        
 [76] cowplot_1.1.3           grid_4.4.1              ape_5.8                
 [79] tidyr_1.3.1             AnnotationDbi_1.66.0    colorspace_2.1-0       
 [82] nlme_3.1-165            patchwork_1.2.0         GenomeInfoDbData_1.2.12
 [85] ggforce_0.4.2           cli_3.6.3               fansi_1.0.6            
 [88] viridisLite_0.4.2       dplyr_1.1.4             gtable_0.3.5           
 [91] yulab.utils_0.1.4       digest_0.6.36           ggplotify_0.1.2        
 [94] ggrepel_0.9.5           farver_2.1.2            memoise_2.0.1          
 [97] htmltools_0.5.8.1       lifecycle_1.0.4         httr_1.4.7             
[100] here_1.0.1              GO.db_3.19.1            bit64_4.0.5            
[103] MASS_7.3-61            
> sessioninfo::session_info()
─ Session info ────────────────────────────────────────────────────────────────────
 setting  value
 version  R version 4.4.1 RC (2024-06-06 r86719)
 os       macOS Sonoma 14.5
 system   aarch64, darwin20
 ui       RStudio
 language (EN)
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       America/Chicago
 date     2024-08-12
 rstudio  2024.07.0-daily+219 Cranberry Hibiscus (desktop)
 pandoc   3.1.11 @ /Applications/RStudio.app/Contents/Resources/app/quarto/bin/tools/aarch64/ (via rmarkdown)

─ Packages ────────────────────────────────────────────────────────────────────────
 ! package          * version date (UTC) lib source
   AnnotationDbi      1.66.0  2024-05-01 [1] Bioconductor 3.19 (R 4.4.0)
   ape                5.8     2024-04-11 [1] CRAN (R 4.4.0)
   aplot              0.2.3   2024-06-17 [1] CRAN (R 4.4.0)
   babelgene          22.9    2022-09-29 [1] CRAN (R 4.4.0)
   Biobase            2.64.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.1)
   BiocGenerics     * 0.50.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocParallel     * 1.38.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   Biostrings       * 2.72.1  2024-06-02 [1] Bioconductor 3.19 (R 4.4.0)
   bit                4.0.5   2022-11-15 [1] CRAN (R 4.4.0)
   bit64              4.0.5   2020-08-30 [1] CRAN (R 4.4.0)
   blob               1.2.4   2023-03-17 [1] CRAN (R 4.4.0)
   cachem             1.1.0   2024-05-16 [1] CRAN (R 4.4.0)
 P cli                3.6.3   2024-06-21 [2] CRAN (R 4.4.0)
   clusterProfiler  * 4.12.1  2024-07-21 [1] Bioconductor 3.19 (R 4.4.1)
   codetools          0.2-20  2024-03-31 [1] CRAN (R 4.4.1)
   colorout         * 1.3-0.2 2024-05-01 [1] Github (jalvesaq/colorout@c6113a2)
   colorspace         2.1-0   2023-01-23 [1] CRAN (R 4.4.0)
   cowplot            1.1.3   2024-01-22 [1] CRAN (R 4.4.0)
   crayon             1.5.3   2024-06-20 [1] CRAN (R 4.4.0)
   data.table       * 1.15.4  2024-03-30 [2] CRAN (R 4.4.0)
   DBI                1.2.3   2024-06-02 [1] CRAN (R 4.4.0)
   default          * 1.0.0   2017-08-07 [1] CRAN (R 4.4.0)
   digest             0.6.36  2024-06-23 [1] CRAN (R 4.4.0)
   DOSE               3.30.1  2024-05-16 [1] Bioconductor 3.19 (R 4.4.0)
   dplyr              1.1.4   2023-11-17 [1] CRAN (R 4.4.0)
   enrichplot         1.24.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   evaluate           0.24.0  2024-06-10 [1] CRAN (R 4.4.0)
   fansi              1.0.6   2023-12-08 [1] CRAN (R 4.4.0)
   farver             2.1.2   2024-05-13 [1] CRAN (R 4.4.0)
   fastmap            1.2.0   2024-05-15 [1] CRAN (R 4.4.0)
   fastmatch          1.1-4   2023-08-18 [1] CRAN (R 4.4.0)
   fgsea            * 1.30.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.1)
   fs                 1.6.4   2024-04-25 [1] CRAN (R 4.4.0)
   generics           0.1.3   2022-07-05 [1] CRAN (R 4.4.0)
   GenomeInfoDb     * 1.40.1  2024-05-24 [1] Bioconductor 3.19 (R 4.4.0)
   GenomeInfoDbData   1.2.12  2024-05-01 [1] Bioconductor
   ggforce            0.4.2   2024-02-19 [1] CRAN (R 4.4.0)
   ggfun              0.1.5   2024-05-28 [1] CRAN (R 4.4.0)
   ggplot2            3.5.1   2024-04-23 [1] CRAN (R 4.4.0)
   ggplotify          0.1.2   2023-08-09 [1] CRAN (R 4.4.0)
   ggraph             2.2.1   2024-03-07 [1] CRAN (R 4.4.0)
   ggrepel            0.9.5   2024-01-10 [1] CRAN (R 4.4.0)
   ggtree             3.12.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   glue               1.7.0   2024-01-09 [1] CRAN (R 4.4.0)
   GO.db              3.19.1  2024-05-01 [1] Bioconductor
   GOSemSim           2.30.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   graphlayouts       1.1.1   2024-03-09 [1] CRAN (R 4.4.0)
   gridExtra          2.3     2017-09-09 [1] CRAN (R 4.4.0)
   gridGraphics       0.5-1   2020-12-13 [1] CRAN (R 4.4.0)
   gson               0.1.0   2023-03-07 [1] CRAN (R 4.4.0)
   gtable             0.3.5   2024-04-22 [1] CRAN (R 4.4.0)
   HDO.db             0.99.1  2024-05-01 [1] Bioconductor
   here               1.0.1   2020-12-13 [1] CRAN (R 4.4.0)
   htmltools          0.5.8.1 2024-04-04 [1] CRAN (R 4.4.0)
   httr               1.4.7   2023-08-15 [1] CRAN (R 4.4.0)
   igraph             2.0.3   2024-03-13 [1] CRAN (R 4.4.0)
   IRanges          * 2.38.1  2024-07-03 [1] Bioconductor 3.19 (R 4.4.1)
   jsonlite           1.8.8   2023-12-04 [1] CRAN (R 4.4.0)
   KEGGREST           1.44.1  2024-06-19 [1] Bioconductor 3.19 (R 4.4.0)
   knitr              1.48    2024-07-07 [1] CRAN (R 4.4.1)
   lattice            0.22-6  2024-03-20 [1] CRAN (R 4.4.1)
   lazyeval           0.2.2   2019-03-15 [1] CRAN (R 4.4.0)
   lifecycle          1.0.4   2023-11-07 [1] CRAN (R 4.4.0)
   magrittr           2.0.3   2022-03-30 [1] CRAN (R 4.4.0)
   MASS               7.3-61  2024-06-13 [1] CRAN (R 4.4.0)
   Matrix             1.7-0   2024-04-26 [1] CRAN (R 4.4.1)
   memoise            2.0.1   2021-11-26 [1] CRAN (R 4.4.0)
   msigdbr          * 7.5.1   2022-03-30 [1] CRAN (R 4.4.0)
   munsell            0.5.1   2024-04-01 [1] CRAN (R 4.4.0)
   nlme               3.1-165 2024-06-06 [1] CRAN (R 4.4.0)
   parallelly       * 1.38.0  2024-07-27 [1] CRAN (R 4.4.0)
   patchwork          1.2.0   2024-01-08 [1] CRAN (R 4.4.0)
   pillar             1.9.0   2023-03-22 [1] CRAN (R 4.4.0)
   pkgconfig          2.0.3   2019-09-22 [1] CRAN (R 4.4.0)
   plyr               1.8.9   2023-10-02 [1] CRAN (R 4.4.0)
   png                0.1-8   2022-11-29 [1] CRAN (R 4.4.0)
   polyclip           1.10-6  2023-09-27 [1] CRAN (R 4.4.0)
   purrr              1.0.2   2023-08-10 [1] CRAN (R 4.4.0)
   qvalue             2.36.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   R6                 2.5.1   2021-08-19 [1] CRAN (R 4.4.0)
   RColorBrewer       1.1-3   2022-04-03 [1] CRAN (R 4.4.0)
   Rcpp               1.0.13  2024-07-17 [1] CRAN (R 4.4.0)
   reshape2           1.4.4   2020-04-09 [1] CRAN (R 4.4.0)
 P rlang            * 1.1.4   2024-06-04 [2] CRAN (R 4.4.1)
   rmarkdown          2.27    2024-05-17 [1] CRAN (R 4.4.0)
   rprojroot          2.0.4   2023-11-05 [1] CRAN (R 4.4.0)
   RSQLite            2.3.7   2024-05-27 [1] CRAN (R 4.4.0)
   rstudioapi         0.16.0  2024-03-24 [1] CRAN (R 4.4.0)
   S4Vectors        * 0.42.1  2024-07-03 [1] Bioconductor 3.19 (R 4.4.1)
   scales             1.3.0   2023-11-28 [1] CRAN (R 4.4.0)
   scatterpie         0.2.3   2024-06-05 [1] CRAN (R 4.4.0)
   sessioninfo        1.2.2   2021-12-06 [1] CRAN (R 4.4.0)
   shadowtext         0.1.3   2024-01-19 [1] CRAN (R 4.4.0)
   stringi            1.8.4   2024-05-06 [1] CRAN (R 4.4.0)
   stringr            1.5.1   2023-11-14 [1] CRAN (R 4.4.0)
   tibble             3.2.1   2023-03-20 [1] CRAN (R 4.4.0)
   tidygraph          1.3.1   2024-01-30 [1] CRAN (R 4.4.0)
   tidyr              1.3.1   2024-01-24 [1] CRAN (R 4.4.0)
   tidyselect         1.2.1   2024-03-11 [1] CRAN (R 4.4.0)
   tidytree           0.4.6   2023-12-12 [1] CRAN (R 4.4.0)
   treeio             1.28.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   tweenr             2.0.3   2024-02-26 [1] CRAN (R 4.4.0)
   UCSC.utils         1.0.0   2024-05-06 [1] Bioconductor 3.19 (R 4.4.0)
   utf8               1.2.4   2023-10-22 [1] CRAN (R 4.4.0)
   vctrs              0.6.5   2023-12-01 [1] CRAN (R 4.4.0)
   viridis            0.6.5   2024-01-29 [1] CRAN (R 4.4.0)
   viridisLite        0.4.2   2023-05-02 [1] CRAN (R 4.4.0)
   withr              3.0.0   2024-01-16 [1] CRAN (R 4.4.0)
   xfun               0.45    2024-06-16 [1] CRAN (R 4.4.0)
   XVector          * 0.44.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   yaml               2.3.9   2024-07-05 [1] CRAN (R 4.4.0)
   yulab.utils        0.1.4   2024-01-28 [1] CRAN (R 4.4.0)
   zlibbioc           1.50.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)

 [1] /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library
 [2] /Users/bmulvey/Library/R/arm64/4.4/library

 P ── Loaded and on-disk path mismatch.

───────────────────────────────────────────────────────────────────────────────────
